requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
flake8==6.0.0
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
numpy==1.24.4
pycodestyle==2.10.0
pyflakes==3.0.1
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
Werkzeug==3.0.1
